requirements.txt
appdirs==1.4.3
awscli==1.11.101
boto==2.47.0
botocore==1.5.64
certifi==2017.4.17
chardet==3.0.4
cloudinary==1.8.0
colorama==0.3.7
diff-match-patch==20121119
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.1
django-ckeditor==5.2.2
django-import-export==0.5.1
django-menu==0.1.9
django-storages==1.5.2
django-tinymce==2.6.0
docutils==0.13.1
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.7.1
idna==2.5
jmespath==0.9.3
mock==2.0.0
olefile==0.44
packaging==16.8
pbr==3.0.1
Pillow==4.1.1
psycopg2==2.7.1
pyasn1==0.2.3
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyuploadcare==2.1.0
PyYAML==3.12
requests==2.17.3
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
static3==0.7.0
tablib==0.11.4
urllib3==1.21.1
whitenoise==3.3.0
